test(color-swatch): recategorize and clean up browser tests
Closes #7184
